requirements.txt
torch
torchvision
faiss-cpu
pillow
numpy
icrawler
git+https://github.com/openai/CLIP.git